+++ .github/workflows/integration-test.yml
@@ -118,8 +118,8 @@ jobs:
       - name: Run trading tests
         run: |
           cd ./examples/js/
-          npx ts-node trade.ts
+          npx ts-node tests/trade.ts
           sleep 5
-          npx ts-node print_orders.ts
+          npx ts-node tests/print_orders.ts
           npx ts-node tests/transfer.ts
           npx ts-node tests/put_batch_orders.ts
